[4.21] Storage: replace artifactory with DataSource in test_golden_image (#4145)
##### Short description:
Manual cherry-pick:
https://github.com/RedHatQE/openshift-virtualization-tests/pull/4095

##### More details:

##### What this PR does / why we need it:

##### Which issue(s) this PR fixes:

##### Special notes for reviewer:

##### jira-ticket:
<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->
