+++ tests/storage/cdi_clone/conftest.py
@@ -1,9 +1,7 @@
 import pytest
-from ocp_resources.data_source import DataSource
 from ocp_resources.datavolume import DataVolume
 
 from tests.storage.utils import create_cirros_dv
-from utilities.constants import OS_FLAVOR_FEDORA
 from utilities.storage import data_volume
 
 
@@ -49,14 +47,4 @@ def data_volume_snapshot_capable_storage_scope_function(
         namespace=namespace,
         storage_class_matrix=storage_class_matrix_snapshot_matrix__function__,
         client=namespace.client,
-    )
-
-
-@pytest.fixture(scope="module")
-def fedora_data_source_scope_module(golden_images_namespace):
-    return DataSource(
-        namespace=golden_images_namespace.name,
-        name=OS_FLAVOR_FEDORA,
-        client=golden_images_namespace.client,
-        ensure_exists=True,
     )
